Add job to sync new IREE fork of StableHLO
We'll use this for carrying local patches for integration purposes as
with our other forks. The other forks all have weird names and aren't
proper forks because they're back from when we were in the google/
GitHub organization and we didn't want to create, e.g.
`google/tensorflow`. Now that we have a separate org, we can just use a
normal fork and that makes syncing much more straightforward. Once we
drop the MHLO repo dep (which this is in service of), I think we can
switch the LLVM fork to be a proper fork and archive the TF and MHLO
ones as no longer used.