Add VmModule.mmap() to Python API.
We really need page aligned flatbuffer blobs vs normal malloc alignment. The best way to be loading a file is via mmap, so just make that available as an API.

This could be done in Python by the caller but is error-prone. The public API will make this more robust.

Provides the mechanism to fix #13887