[CPU] Scalarize vector loads used in a scalar fashion
This PR makes sure that we scalarize the vector load that are used in a scalar
fashion (i.e., vector.transfer_read + vector.extract) in matmuls. The LLVM
backend is sometimes able to scalarize this vector load but not always.
For cases where it's not scalarized, register pressure increases as the
liveness of this vector load spans across all the unrolled iterations of
the matmul. There are also some shuffles/permutations generated that don't
make too much sense from the performance point of view.