Fix the ukernels system build on older toolchains. (#14155)
In the process of adding the ukernels bitcode build, we dropped all
cmake configure-check for toolchain support for CPU-feature-enabling
flags, and configured headers. I didn't properly think through that:
that worked essentially because no one had tried building with an older
toolchain. On x86-64, that was OK because we didn't use any recent flag.
but on ARM that was more problematic with the `+i8mm` target feature.
@freddan80 ran into this on
https://github.com/openxla/iree/issues/12684.

So this brings back configure-checks and configured-headers, but only
where they are specifically needed and not interfering with the bitcode
build --- only in the arch/ subdirs and only in the system-build. Some
`#if defined(IREE_DEVICE_STANDALONE)` lets the bitcode build opt out of
including the configured headers.

This has a couple of side benefits. We get to drop the clumsy `#if`'s
trying to do version checks on compiler version tokens, and we get to
conditionally add those feature-specific dependencies instead of having
those `#if`'s around the entire files, which confuse syntax highlighting
when the feature token is not defined in the IDE build.