Delete old e2e tests with massive checked-in IR files
These were mostly added years ago before we had mechanisms for testing
models end-to-end. The tests here don't use our standard test
infrastructure and instead JIT-compile with `iree-run-mlir`. This is
nice to provide some small sample programs that can be iterated on
without having to think about build systems, but is not appropriate for
massive IR files. Now we have a better way. There are already
(CMake-only) end-to-end tests
[in this directory](tests/e2e/models/generated_e2e_model_tests.cmake)
that compile using the build system with the accompanying caching,
dependency analysis and parallelism and then the test is only actually
executing the module. The immediate motivation here is to delete the
resnet50 test that is our slowest on every platform and keeps timing out
or needing to be disabled.