Switch compiler/ to LLVM/MLIR formatting style
The high-level goal is to better align with the LLVM/MLIR coding style
and allow for easier prototyping and code movement between the two.

This adds a new `.clang-format` file local to the `compiler/`
directory with the same formatting settings as those used by MLIR. It
also reformats this directory using the new formatting config.

I used the following command to reformat the code:
```shell
  cd compiler
  fd -e h -e cpp -e cc -e c | xargs clang-format -i
```

Fixes: https://github.com/openxla/iree/issues/12866