Relax the Matmul transform strategy to allow using fmas for smaller cases.
This PR generalizes the matmul strategy to degrade to non-tensorcore operations.
This is particularly useful for matmuls with some small dimension(s) that can
benefit from more aggressive schedule of copies that don't match tensorcore operations.

This was tested to high fraction of bandwidth peak for such cases of small dimension(s).