Switch compiler to LLVM/MLIR formatting style (#14181)
The high-level goal is to better align with the LLVM/MLIR coding style
and allow for easier prototyping and code movement between the two. The
`runtime/` and `tools/` directories still use the `Google` style, as the
LLVM config is not always preferred for C code, and we do not expect
much sharing/movement of this code with LLVM.

This adds a new `.clang-format` file local to the `compiler/` directory
with the same formatting settings as those used by MLIR. The only
divergence is not having special case for include ordering that places
llvm includes last. It also reformats this directory using the new
formatting config.

I used the following command to reformat the code:
```shell
  cd compiler
  fd -e h -e cpp -e cc -e c | xargs clang-format -i
```

I plan to add this commit to `.git-blame-ignore-revs` in a follow-up PR,
once the final commit SHA is known.

Issue: https://github.com/openxla/iree/issues/12866