Add support for lowering ops that implement the `UKernelOpInterface` on the CUDA path. (#14101)
This change add support for lowering operations that implement the
`UKernelOpInterface` on the CUDA backend. This allows custom
dispatches to use ops that implement and be compiled down to CUDA.

Unlike the CPU side, the CUDA side needs more information than just
the number of workgroups to use. It also needs the workgroup size. To
allow specifying this, an #iree_codegen.export_config attribute is
added. This allows specifying the workgroup size to use on the custom
dispatch. This information is used to set the workgroup size to use
during `MaterializeInterfaces` in HAL.

An example of the bitcode usage is added with this PR. Since there are
no in-tree examples of CUDA micro-kernels, the bit code is linked
using the `iree-link-bitcode` (which is the
`iree-llvmcpu-link-bitcode` renamed).

Fixes #14028