Make barrier elimination more aggressive (#14293)
Contrary to the conceptual description of the barrier elimination
technique in the paper by Moses et.al., barrier elimination in IREE is
performed by the greedy rewriter. Thus a redundant barrier is erased
immediately and the updated IR is considered after for the next barrier.
Therefore, there it is sufficient to consider the effects before and
after each barrier until hitting the next barrier, rather than pairwise
extending those until the end of the parallel region boundary.

Specifically, the case that wouldn't be handled correctly with non-eager
rewriting,

```mlir
  store %A
  barrier  // useless because no effects after
  // nothing
  barrier  // useless because no effects before
  load %A
```

is handled correctly in IREE that removes one of the barriers and sees
the second as required when re-analyzing the eagerly rewrtitten IR.

This change lets us eliminate more barriers.