Makes --iree-input-type an open namespace and adds a plugin extension for it.
This lets plugins define their own input types. The plumbing on the plugin manager side is a bit awkward since typical pipeline construction constructors don't have access to the context for error reporting. We may need to revisit a better error reporting option later if this gets any more complicated. However, there is only one use, so it is unlikely to spider.