Make the MLIRInterop.h APIs more ergonomic.
The previous APIs didn't really have a way to manage context or module lifetimes that may conflict with what the caller needs.

Now we have these APIs:

* ireeCompilerSessionBorrowContext
* ireeCompilerSessionStealContext
* ireeCompilerInvocationImportBorrowModule
* ireeCompilerInvocationImportStealModule

Among other things, this allows us to transfer ownership of context/modules in a way that is compatible with the MLIR Python API (which generally must own everything it sees).