Reland "Implement CI job optionality (#14312)" (#14362)
I tried to be consistent about json-encoding everything in `set_output`,
so the caller didn't need to worry about that, but it turns out that
wraps things that are already strings in an extra set of quotes,
breaking ~everything.

Additionally, I wrote a function with two positional boolean args
(bad!) and swapped the order they were passed (which is why those are
bad). Changed them to keyword args.