Adds support for serialization of sub-byte aligned element types. This currently assumes (poorly) that physical storage is always the next power-of-two byte width greater than 2x the logical bit width, so i3 is packed into 2xi3 in i8 (2 wasted bits), i5 is packed into 3xi5 in i16 (1 wasted bit), and i9 is packed into 3xi9 in i32 (5 wasted bits).
When we get proper packed types that let us express the physical and
logical bit widths independently we can change the
getRoundedPhysicalStorageSize method to use it and most of the code should
still work. There's performance improvements to be had when serializing
out large arrays of values but that can wait until we know the proper
physical sizes.