[WIP] Remove `tensor.empty` for fusion cases.
For a dispatch of the form

`tensor.empty` -> `linalg.fill` -> `linalg.matmul` -> `linalg.generic`

in some cases, the `tensor.empty` can be replaced with using the
output of the `linalg.generic` directly if the result of the
`linalg.matmul` has the same type as the result of
`linalg.generic`. This transformation already existed in the
`ConvertToDestinationPassingStyle` pass but was removed since the
assumption was that after vectorization the `tensor.empty` would be
dead. Adding this back since for non-vectorization cases this will
result in an unnecessary stack allocation.

Fixes #14316, #14305