[CPU] Improve vector size computation in Vectorization pass
This PR should be mostly RFC for existing working cases. It simplifies
the retrieval of the vector sizes by removing the `getCanonicalVectorShape`
and only looking at the lowering config of the operation to be
vectorized and no longer looking at the lowering config of another (root)
operation. It also gives priority to the lowering config found in the operation
to be vectorize and fall back to retrieving it from the IR when the lowering
config is not found.