Link to website overview slides in docs/README.md. (#14400)
I presented these slides at an internal team meeting and may repeat the
presentation for the broader community if there is interest and a time
slot.

The slides should be world-readable (except where `docs.google.com` is
blocked). I tried exporting to .pdf but don't want to commit a 1.3MB
file to the repo (lots of screenshots, even with good compression).