Add option to initiate compilation from a specified phase
Compiler re-entry can aid in debugging and hand manipulation of IR at
various phases of compilation flow without having to resort to
`iree-opt` commands. This also helps with applying code generation
configs without having to go back through the Flow passes, or with
skipping earlier phases of the pipeline if they are taking up most of
the compilation time.

This adds a `compile-from` flag to initiate compilation at (including)
the specified pipeline phase, similar to the existing `compile-to` flag.