[TransformStrategies] Add implicit gemm for aligned convolutions in CUDA (#13449)
This adds a builder for implicit gemm based closely on the existing
strategy for aligned matmuls. The strategy aims to be as close to the
matmul strategy as possible, only deviating when it comes to the
handling of the img2col op for forming the column matrix in shared
memory. There are a few additional restrictions:

1. Due to the img2col op, padding on the convolution input image will
   fail, so currently we restrict to tile aligned cases.
2. Currently restricted to 2d named convolutions (nchw_fchw and
   nhwc_hwcf). Further convolution variants will be added with
   generalizations to img2col.
3. Async copies and multibuffering + pipelining currently fail with the
   img2col op and are disabled.

To match the existing matmul strategy, as well as due to these
restrictions, the strategy is currently locked behind a flag

`--iree-codegen-llvmgpu-enable-transform-dialect-implicit-gemm-strategy`.