Adding Vulkan sparse binding buffer support for native allocations. This creates one logical VkBuffer that is backed by as many aligned max-size allocations as required. There's a lot we could tweak here and a lot to optimize but the initial proof of concept here is specifically for allowing large constant/variable buffers with long lifetimes.
Future changes will add asynchronous binding and sparse residency as
part of the HAL API so that targets supporting constrained virtual memory
management (CPU, CUDA, Vulkan, etc) can have such virtual/physical
remapping exposed for use by the compiler. When that's implemented the
sparse buffer type here will be reworked as a shared utility
implementation using the binding/sparse residency APIs.

In order for this to be used for large constants host allocation
importing was implemented so that the buffers can be transferred. This
required a change in the HAL APIs exposed to the compiler as what was
there was a hack to approximate the proper import/mapping path but
insufficient for doing it properly.

Progress on #14607.
Fixes #7242.