[CPU] Fix tile sizes adjustment for pack op with outer_dims_perm. (#14441)
The tile sizes of distribution and TileAndFuse are set based on rootOp.
They should be transposed if outer_dims_perm exists. Otherwise, they
won't be applied to corresponding loop. E.g., say that the expected tile
sizes for root op are `[8, 128]` and the pack op consumer has
outer_dims_perm. If they are not transposed, below snippet would be
generated and a `128x8` slice of rootOp would be created.

```
%producer = linalg.generic
scf.for ... step 8
  scf.for ... step 128
    %slice = tenser.extract_slice %producer [%i, %j] [128, 8] [1, 1]
    %pack = tensor.pack %slice outer_dims_perm = [1, 0] ...
```

What we want is to create a `8x128` slice. So we have to transpose tile
sizes if outer_dims_perm presents.

Fixes https://github.com/openxla/iree/issues/14252