[CPU] Add `contract` fast-math-flag to arith operations
This patch adds the `contract` FMF to some arith operations so that they can
be folded into an fma instruction. We are doing this by default as we
are lowering matmul ops by default to fmas. We will add different fp
modes to have more control on fp optimizations depending on the
tolerance to fp errors.