Isolate LLVMTarget in the llvm-cpu backend. (#14516)
Isolate LLVMTarget in the llvm-cpu backend.

This carefully pry's apart the LLVMTargetOptions from the LLVMTarget,
making sure that the latter is hermetic for anything affecting the
generated code.

Further replumbs the target backend so that it doesn't reach into
default state and works off of the configuration IR exclusively.

Also updates the consteval JIT to use this new capability for targeting
the host, even if the overall compilation is cross-compiling.

Flags changed:

* Removes `iree-consteval-jit-use-vmvx`
* Adds `iree-consteval-jit-target-backend`: Currently still defaults to
VMVX but with this change it should be possible to reliably flip it to
llvm-cpu.

I tested this by adding a similar print function to the options in a
base branch and exhaustively compared the results of of the target
information before and after the change for full builds. There were some
tricky things in the original since it was initializing flags off of
defaults which would then be re-calculated and such. I've smoothed all
of this out so that at least the paths are straight through once the
initial configuration is set.