Switch consteval to use llvm-cpu by default.
This is possible because #14516 makes it possible to recursively target LLVM-CPU.

Also added a check on the env var `IREE_COMPILER_DEBUG_CONSTEVAL` to enable the debugging mode for consteval because these kinds of compiler-in-compiler things are murderous to debug if needing to trace through the whole thing and set a flag. I'll probably generalize this to a more general "IREE_COMPILER_DEBUG" flag that we can use compiler-wide for in the field debugging output (later).