Start pipecleaning DenseElementsAttr direct references.
We will be making more use of ElementsAttr (the interface) and we need to avoid direct assumptions on DenseElementsAttr.

In order to improve this, I have added a ManagedElementsAttrInterface and have added methods there that are needed to mutate and do other things. I will need to extend this later to more efficiently manage the resources, so funneling it all through here helps.