Add canonicalizer that reorders unary elementwise ops and shape manipulation ops. (#14494)
The pass identifies shape manipulation operations (`reshape`,
`transpose`, `broadcast`) that feed into unary elementwise operations
and swaps their order. Since elementwise ops do not manipulate the input
shape (and there is no interaction between the elements), the reordering
doesn't affect the result.
This enables a future PR where some consecutive shape manipulation
operations should be collapsed together into a single operation.
`Broadcast_in_dim` is not included in this PR as the same thing is done
for it as part of an existing `StablehloToStablehlo` pass.
In the future, we might want to consider merging/more clearly defining
the difference between `Canonicalizations.cpp` and
`StablehloToStablehlo.cpp` files, as at the moment their function is
almost indistinguishable in some cases.