Add a preprocessing pass to move entire function into a single dispatch.
For cases where the model is very small and does not have much
concurrency, it is better to move the entire function body into a
single dispatch. Eventually the default heuristics can probably figure
out when a model is "too small", but for now this PR adds a pass to
move the entire function body into a single dispatch to use as a way
to find codegen issues such an approach throws up, and also to
experiment with different heuristics needed to find such dispatches
automatically.