[gpu] Distribute fused producer elementwise ops in SIMT pipeline (#13908)
The fused producer ops have different dimensions than the consumer
matmul op. Tiling and distribution the matmul op's configuration may not
the the best.

This commit postpones the producer op tiling distribution to where copy
ops are handled. There we use all threads in the workgroup to distribute
in a flat manner.