Build the Tracy-enable iree-runtime package by default.
This has been stable for the last week on CI/interactive use. Flipping to default as it is always supported.

`IREE_RUNTIME_BUILD_TRACY_TOOLS=ON` remains opt-in because that requires system dependencies to be met (this gates the availability of iree-tracy-capture). It is manually enabled in CI on platforms that support it.