Restrict the constexpr constant types to the subset we currently support (#14758)
Notably, this eliminates constexpr identification for non-byte aligned
types (except for i1, which is presently stored in i8) that are not
supported by the JIT. Note that the analysis for these seemed wrong
anyway, since it was possible to hoist an expansion. As all of these
types need work, we just opt them all out for the moment.

This also fixed a bug where the type of the root constant node was not
being checked for legality.