Move demotion passes to GlobalOptimization.
The global optimizations really depend on the flow level demotion passes. Moves them to the right place. Also restores the ordering of the strip assertions pass so that it can guide optimizations.

Renames old `-iree-flow-(demote|promote)-*` flags to `-iree-opt-(demote|promote)-*` flags now that they are just part of the rest of the global optimizations.

Internally, renames "HighLevelOptimizations" to "GlobalOptimizations" for coherence.

Fixes #14835.