Move demotion passes to GlobalOptimization. (#14815)
The global optimizations really depend on the flow level demotion
passes. Moves them to the right place. Also restores the ordering of the
strip assertions pass so that it can guide optimizations.
    
Renames old `-iree-flow-(demote|promote)-*` flags to
`-iree-opt-(demote|promote)-*` flags now that they are just part of the
rest of the global optimizations.
    
Internally, renames "HighLevelOptimizations" to "GlobalOptimizations"
for coherence.

This lets us enable consteval on llam2 7b qi4/f16 models and drops
latency from 14.3ms -> 11.9ms. Mostly it is small scalar-level
evaluations but it is also eliminating a 250MiB f16 transpose in both
first/second that I expect is a main culprit.

Fixes #14835.