[cuda] Fix crash caused by CUevent outliving CUdevice
We use `iree_hal_cuda2_event_t` to wrap a `CUevent` so that we can
have multiple places referencing the `CUevent` with automatic
resource releasing when refcount becomes zero.

The `iree_hal_cuda2_event_t` objects are further placed in a pool
to amortize the cost of creation via recycling. But how it is
set up right now, the pool's lifetime is associated with the
HAL device. So it will be destroyed together with the device,
including the CUDA context and so. All events that are in the pool
will be destroyed; but there is no proper handling of outstanding
`iree_hal_cuda2_event_t` in this flow.

If a `iree_hal_cuda2_event_t`'s refcount decreases to zero after
device destruction, it will cause segfaults in CUDA driver given
the CUDA context is already gone.

So this commit fixes the issue by retaining the pool in the
`iree_hal_cuda2_event_t` object and retaining the HAL device
in the pool.