Support multi-block functions in EraseHALDescriptorTypeFormMemRef (#14879)
If we have functions with multiple basic blocks and the HAL descriptor is part of a block signature, we have to replace the function when getting rid of these descriptors. This PR adds the logic to remove these descriptors and makes sure the pass runs from the function's parent op (otherwise, we couldn't replace the function as part of the transformation).