Remove redundant parameters from ukernel tile funcs (#14895)
ukernel tile funcs used to take by value some params fields, even though
they also took the params pointer from which they could be obtained.

This trait dates back to when tile funcs were implemented in out-of-line
assembly. Passing by value the two fields that they needed avoided
having to worry about using the right field offsets in assembly. We
don't have any out-of-line assembly anymore as that was incompatible
with compilation to bitcode, so we don't need that anymore.