Replace tm_tensor with torch-mlir proper (#14917)
Per discussion on iree-discuss, this patch removes the original
tm_tensor source snapshots and special casing from the codebase, instead
pulling in the parts of torch-mlir that provide a full solution,
starting from the torch dialect down.

This reuses the existing `IREE_INPUT_TORCH` CMake define but routes it
to a compiler plugin at `compiler/plugins/input/Torch` where all build
plumbing to integrate and local pipelines exist. I have kept the
original tm_tensor input type for the moment since downstreams are using
that, but we will want to eventually replace that with a dedicated
`torch` input type which works directly from the `torch` dialect.

This patch should be NFC to all current users, however it allows
programmatic and API access to the full Torch dialects and conversion
pipelines, which is essential for broader scale integration.

We add `third_party/torch-mlir` as a submodule at torch-mlir's current
HEAD. During LLVM integrates, if there are ever API breaks that affect
this, we may need to push patches to torch-mlir to address. This can be
done on a torch-mlir branch as needed (or committed at head). I have
been tracking changes for a couple of months and have not had a need to
do this yet, which indicates that this is relatively stable.

Committing with an XFAIL'd lit test, which for some reason was not
running previously and now is:
https://github.com/openxla/iree/issues/14916