Refactor IREECodegenAttrs to use typed array parameters
This reworks the attributes in IREECodegenAttrs so that typed
[Optional]ArrayRefParameters are used instead of untyped ArrayAttrs.

Doing this removes the manual type-checking and manual conversions
to/from SmallVector<T>.

This did require a small restructuring of the lowering_config attribute,
so instead of having multiple nested arrays IntegerAttrs for tiling
levels, it now has one array of TilingLevelAttrs. The TilingLevelAttr
then contains the tile sizes and tile interchange values for a
particular level.

The TilingLevelAttr has a custom assembly format that keeps its syntax
as close as possible to the previous syntax, to avoid as much test
churn as possible. We plan to follow up this patch with another change
that extends the TilingLevelAttr to support scalable sizes (using the
canonical MLIR syntax), which will need a custom parser anyway, so we
don't see this as a dealbreaker.

Example configs:

```
  // No tile interchange
  // Before:
  #iree_codegen.lowering_config<tile_sizes    = [[2, 64, 64, 64],
                                                 [1, 1, 1, 4],
                                                 [0, 0, 0, 0]]>
  // After:
  #iree_codegen.lowering_config<tiling_levels = [[2, 64, 64, 64],
                                                 [1, 1, 1, 4],
                                                 [0, 0, 0, 0]]>

  // With tile interchange
  // Before:
  #iree_codegen.lowering_config<
   tile_sizes = [[32, 64, 0], [8, 32, 0], [0, 0, 16]],
   tile_interchange = [[1, 0, 2], [], []]>

  // After:
  #iree_codegen.lowering_config<tiling_levels = [
   {sizes=[32, 64, 0], interchange=[1, 0, 2]}, [8, 32, 0], [0, 0, 16]]>
```