[spirv] Fix transfer op crash when vectorizing memref
In the `SPIRVVectorizeLoadStore` pass, we try to use the smallest
vector as the new element type for vectorizing scalar memref types.
This means afterwards we can see transfer ops where we are reading
or writing a larger vector size, e.g., vector<4xf32> -> memref of
vector<2xf32>. Such cases would require us to unroll the vector
and perform read/write for its subparts.