Add a dedicated 'torch' input pipeline. (#15037)
This has been in use downstream via the `torch-to-iree` pass and is now
ready to be exposed as a real input type.

As of https://github.com/nod-ai/SHARK-Turbine/pull/72, Turbine can now
generate pure 'torch' level programs that can be fed to IREE directly
with this input pipeline. We expect that this will be more stable and
supportable vs half-lowering to IREE internal dialects in the frontend.