Add a way to hook custom TD strategy for specific ops
This patch adds a transform dialect interpreter pass that can be used to
annotate specific operations with specific strategies. This patch relies on
https://github.com/openxla/iree/pull/14788 to actually "link" the strategy
within the related module.

The intended use case, as demonstrated in the added test cases, is to:
1. specify the matcher in a dedicated file (in the transform dialect format)
   that is passed to the compiler through
   `--iree-llvmcpu-transform-dialect-select-strategy`.
2. provide the strategy as a named sequence through the library option
   `--iree-codegen-transform-library-file-name`.

If the matcher applies in #1, then the transform dialect pipeline will pick
up the proper strategy for #2 and apply it to the annotated operations.