Restore PJRT to minimally functional after break with XLA.
When imported, we removed all code that relied on static/bazel linkage into XLA internals. This patch adds back the pure C path for device memory layout, which is the minimum needed to run simple programs.

Next, we will need to add code generation for layout changes, and then it is just minor ergonomic work to get back to where it was. As-is, though, we can run things with this patch again and can start wiring it for CI, etc.