Splitting dispatch annotation from outlining. (#15177)
It was always a hack to have the annotations happen during outlining and
it's since grown unweildy. Separating also has the benefit of allowing
executables that come in from frontends (not outlined/formed by us)
being able to be named based on their contents.

Note that executable names now don't change which is intentional:
executables are not intended to contain a single dispatch and on CPU
there's no difference as it already links multiple dispatches into the
same executables. CUDA/SPIR-V/etc all need to link dispatches together
and maybe this will encourage that ;)