Fix allocator crashes on Apple (#15199)
Some tests are crashing on macOS. `std::aligned_alloc` is being
incorrectly used. Two issues:

1. The alignment<->size parameter order was reversed.
2. Unlike other similar aligned-alloc functions, it requires the size to
be a multiple of alignment. So we have to do the rounding up to next
multiple of aligment on our side. And then, since we should try to be
correct in allocator code, we have to be careful with the possibility
that that would overflow the size type.