Adding hal.dispatch.extern op. (#15193)
This is an inline dispatch op on tensors that can be inserted in source
programs to fully specify an externally derived dispatch region.
Improvements were made for external executables in flow/stream/hal.

There's a decent amount of cruft on these paths that'll need to be
cleaned up; particularly MaterializeInterfaces has long been in need of
a reworking (likely decomposing with the assistance of a new op or two)
and it's not great that flow now has handling of a HAL pseudo-op. The
stream convert HAL->stream which handles hal.tensor.import/export would
be a better place but executable deduplication happens in flow today and
we need inlined dispatches like this to deduplicate. I toyed with also
having a DeduplicateExecutables in stream but decided against it for
today.