Add support for external custom dispatch rewrite patterns
Currently custom dispatches must be written in as a part of downstream
flows (targeting `hal.dispatch.extern` or Flow ops directly). This
limits how lightweight deployment of custom dispatches with nearly stock
IREE flows. This adds a mechanism to plugin and apply pdl patterns
completely separate from other IR. This allows hermetically sealing the
details of a custom dispatch as a separate IR module that is easy to
reuse across models.

The intent is to follow up to this patch with similar passes in the
various input dialects to allow different potential rewrite points, as
well as turn some of the rewriting utilities into transform operations
as another way to represent the dispatch + rewrite logic without the
need for C++ and a downstream project.