Improving the `*.tensor.trace` op to carry shapes/encodings. This avoids a bunch of IR noise when lowering into the stream dialect and forces all traced tensors to be transfered into staging resources. This means we no longer require mapping (or worse, emulated mapping) on any target in order to trace as we'll have the compiler implement the readback asynchronously.
A change I held off on here but should be done soon is making the trace
ops return their tensors so that we can asynchronously schedule them.
Then instead of emitting runtime calls on staging resources we could
batch up the transfers into a ringbuffer and trace out without
introducing dispatch-to-dispatch host synchronization.