Making execution region results queue-ordered allocas. (#15149)
We don't currently insert deallocas and don't track live ranges but that
can come in the future as we support more control flow. For now this at
least gets all of the common allocations within an invocation into the
queue-ordered bucket so that we can do proper async execution and use
native queue-ordered (e.g. stream-ordered allocations in CUDA)
functionality.

With this change the caching allocator is no longer needed for CUDA in
almost all cases (besides exported function results).