[Flow] Add TensorBitCastOp
This is a pseudo-rfc about how to handle bitcasting of tensors within
IREE. This patch adds `flow.tensor.bitcast` as a near mirror of
`flow.tensor.reshape`, however allowing changing element type bit widths
as well. This allows earlier lowerings to skip materializing a constant
tensor of some difficult to represent type and instead bitcast from
a nicer byte-aligned and/or integer type. Similarly, this can help
bridge the gap between frameworks, which might have limited support even
for integers of varying bit widths, and IREE.

In terms of direct applications today, this removes the need to
materialize the sub-byte constant tensors for quantized LLMs like LLaMa
that have seen recent burn-downs. As a result, we can store the
constants as i8 instead of converting the elements one-by-one to APInt
to allow MLIR to represent the constant tensor, and instead just keep
the values as is from the frontend (the frontend is giving it to us
packed!). This should improve memory usage at compile time by at least
a factor of 2 (or more, not sure how APInt is storing those values), as
well as give significant compile time gains both at load and
serialization time.

If this op is too much of a clone of reshape, it's easy to just simplify
the verifier for reshape, but I figured letting the reshape also bitcast
(with changing bit width) might be doing too much, and I see this
bitcast op as introducing potential footguns/points of ambiguity.
Essentially with this op, we would have a situation where the storage of
a non-power-of-two sub-byte resource could be ambiguous from codegen's
perspective.

Currently let's say we have a constant of `tensor<64xi3>`. Currently
this would be serialized such that 2 `i3`s are packed per byte with
2 wasted bits. If we instead had a constant of `tensor<24xi8>` and
bitcasted to `tensor<64xi3>`, semantically a bitcast is a no-op and thus
the resulting `i3` tensor will be stored with 24 bytes as opposed to 32.
Codegen would only see the interface binding + offset for
a `tensor<64xi3>` and thus can't know how to generate code.

(see compiler/src/iree/compiler/Dialect/VM/Target/Bytecode/test/constant_encoding.mlir)

I haven't thought through the right way to handle this yet, but I also
figured that codegen can't really handle non-power-of-two types right
now anyway so I ignored the problem for now :/