[AArch64][SVE] Add e2e tests for small and large f32 matmuls
This requires #15276, which prevents IREE from generating scalable
vector types that are poorly supported by the LLVM backend (and will
likely result in it aborting).

These tests are tagged as requiring aarch64+sve, so currently won't be
run in CI (as there's no SVE capable runners yet). When run locally (on
a machine with SVE) these tests pass, and do use scalable vectorization.