[Codegen] Add ability to specify transform dialect libraries (#14788)
This adds the ability to specify transform dialect strategies through
a symbol pointing to a library call. This is currently available through
two flags:

`--iree-codegen-use-transform-dialect-strategy=[@<symbol_name>|filename]`
  `--iree-codegen-transform-dialect-library=filename`

The transform library is loaded and cached in the IREE codegen dialect
for subsequent invocations within the MaterializeUserConfigs pass. Then,
the loaded dialect is immediately used with the symbol name referenced
by the transform dialect usage flag. If a filename is specified instead,
that is broadcasted to the transform dialect interpreter (intended for
microbenchmarking).

If the symbol applies successfully, this will send the result through
normal IREE codegen. This gives users the option to replace the
translation info on the export op with a `<None>` pipeline to send it
through <BACKEND>LowerExecutableTarget unperterbed (thereby skipping
the initial tile + distribute and bufferization).

Additionally this unifies the way the transform dialect testing flags
currently duplicated across backends.