Making flow.dispatch/stream.async.dispatch take multiple symbols. (#15295)
stream.cmd.dispatch already supported this for making external
hal.executable.variant ops work and by making this consistent up the
stack it allows for the use of hal.executable.variant all the way up in
flow. This will allow hal.dispatch.extern to expand to HAL ops instead
of flow.executable and avoid the need for plumbing all of the
HAL-specific behavior through those layers.