[Vulkan] Make vulkan target triples/envs lists
This makes it easier to specify multiple vulkan targets simultaneously.
If there are target triples interleaved with target environments, the
order of targets is inferred from the order of flags.

As a follow up, we can do the same to hal device targets, allowing one
to specify the full target attribute from the command line or some other
downstream project.