[NFC-ish] Split out kernel config into <BACKEND>SelectLoweringStrategy (#15321)
This slightly restructures codegen in all of the backends such that the
selection of the lowering strategy is separate from the construction and
application of the relevant pipeline. This has two potential benefits:

1) By decoupling strategy selection from the lowering, we can move the
translation info onto function entry points (and/or individual private
functions) rather than the export op. This allows codegen to nest on the
function rather than the variant, whereas before this change, strategy
selection would still need the higher level of nesting.

2) We can now insert passes between strategy selection and application.
This allows expansion of variants based on the set of required
capabilities for each strategy, and contraction across variants that can
now be statically resolved as identical (think an elementwise operation
being independent of SubgroupNonUniform, now we can collapse two
variants where the presence of SubgroupNonUniform was the only thing
present).

Unfortunately for the time being, we still need to register essentially
the same set of dialects during the strategy selection pass because the
Transform Dialect builders in TransformStrategies can more or less
produce whatever they want. Once those strategies are deprecated we can
drop the extra registrations.