Fix size calculation in the tensor.empty materialization pattern.
The result of `getDyanmicSizes()` and shape sizes mismatch if the
original input is partially dynamic. E.g., the former one can return
"%c1, %x, %c4096", while the latter one can return "1, DYNAMIC, 4096".
They mismatch in `getMixedValues()` method because the dynamicValues
should only contain `%x` element.

Since that we have the original information in the encoding, we can
assume that they are inferred to static values during setting encodings.
We can fold empty op's operands to attributes if they are constants. It
matches all the information we have in MaterializeEncodingInfo, and
`PackOp::getResultShape` can infer the actual type.