Cast tensor.empty type to TypeConverter's type during materialization. (#15375)
It is a bug introduced by
https://github.com/openxla/iree/commit/03d655a99408f43bd5d960172b15fb0d5823e633

VMVX has a different assumption because it generally supports dynamic
cases. I.e., it treats all the matmul as dynamic shapes even they are
static inputs. This could make inferred shape and TypeConverter's target
shape different in terms of dynamism. Casting it to the target type to
make dialect conversion happy.