[SPIRV] Handle create_mask lowering during load/store vectorization
Proper handling of masks requires direct lowerings of the mask, however
we can skip the materialization of the mask entirely (and subsequent
reliance on unrolling/canonicalization to clean up the masks) by folding
`vector.extract(vector.create_mask)` into the appropriate boolean. This
works for subgroup reduce (the first dynamic codegen problem we're
tackling) because masks and transfers are never unrolled in this pipeline,
rather we just distribute them in place to a pre-configured vector size.

We handle this during SPIRVVectorizeLoadStore because immediately after
scalarizing vector.transfer_read/write, we will form the extracts on the
mask.