Generalization for ElementsAttr coverage (#15433)
This PR addresses the issue that there was code written in IREE from a
time when there was only DenseElementsAttr. DenseElementsAttr is a
specific type and ElementsAttr the interface. There are actually
multiple implementations of the interface now and DenseElementsAttr is
being phased out for general usage. This PR deals with generalization,
so we are not boxed into only supporting DenseElementsAttr.