Add dep for ukernel internal headers (#15462)
The Bazel to CMake converter currently ignore internal headers Bazel
dependencies. This means that incremental build does not rebuild
ukernels upon changes to a header file which can lead to subtle ABI
issues.

The patch now uses ${CMAKE_SOURCE_DIR} or ${CMAKE_BINARY_DIR} when
converting a package-relative label to a path because DEPENDS in a
add_custom_command() resolve either relative to the
${CMAKE_CURRENT_BINARY_DIR} or ${CMAKE_CURRENT_SOURCE_DIR} whereas
reference in other Bazel package are relative to the root of the tree.