[GlobalOptimization] Lift `linalg.generic` ops to `linalg.batch_matmul/linalg.batch_vecmat/linalg.batch_matvec` (#15339)
This adds a pass that lifts generalized batch matmul `linalg.generic` ops 
into the corresponding linalg name ops. Any element type promotion in 
the `linalg.generic` op is rewritten as a producer op, and the element 
types of the new linalg named op will be the already-promoted type. This 
allows having ukernels with unsigned types, since the unsignedness of 
the inputs can be inferred from the producer op.

---------

Co-authored-by: Benoit Jacob <jacob.benoit.1@gmail.com>