Resolve `host` CPU to `generic` outside of x86 (#15481)
Following https://github.com/openxla/iree/pull/15477, `cpu` is treated
as a x86-only thing, except for the special value `host` that gets
resolved to the actual host CPU and CPU features.

JitGlobals creates a host target, that gets resolved accordingly. The
resolved CPU is then recorded in a `hal.devices.targets` attribute. If
the CPU field is set to the actual host CPU identifier there, then that
will trigger the error message about that being a x86-only thing:

```
error: Resolution of target CPU to target CPU features is not implemented on this target architecture. Pass explicit CPU features instead of a CPU on this architecture, or implement that.
```

So this needs to be left empty in that case. No functional
difference, as that CPU is just ignored anyway (which is what the error
message is about).