Setting worker local memory size based on L2 cache size by default. (#15444)
Users can override this with --task_worker_local_memory= as before (or
when programmatically creating the executors) but on systems where we
have cpuinfo or Windows we'll pick it up automatically. Mac, bare-metal,
etc will need to pass the flag for now until topology query support is
implemented for them.

Fixes #15440.