Add hooks in const expr hoisting to keep sub-byte tensors packed
Non-byte aligned constant tensors are stored as an array of APInts,
meaning it consumes multiple times as much memory and requires repacking
later on during serialization. This has a significant impact on compiler
performance for models that use such types.

To avoid this problem, this adds a hook to const-expr hoisting that
allows setting a preferred storage type for globals, along side
callbacks to materialize casts to and from the desired type. We do this
with pass options instead of directly in the pass itself to avoid Util
from taking a dependency on Flow.