[SPIRV] Enable dynamic reduction through subgroup reduce (#15546)
This only allows single dynamic reduction dims through subgroup reduce
for now because support for multi-dim masked write distribution is a bit
hazy. Also, in the absence of better information for the expected size
ranges for those dynamic dims, this keeps to using a single subgroup and
vector size of 1 because all reads/writes will be scalarized and
predicated anyway.

There can be follow up work to start peeling off the dynamic portion
of the local reduce, thereby allowing efficient vectorized
loads/stores/compute for the majority of it, and only
masking for the last iteration.

Also, the current logic for picking the masked vector sizes to use
relies on logic in TilingConfig which is incompatible with other
backends. For now this just adds an option to skip inference of tile
sizes, but in the future we will likely have to decouple the CPU centric
vector size inference logic from GenericVectorization.