Add optional CMake presets to the repo. (#15581)
Since people have complained about this in the past with respect to IDEs
doing weird things, I created these in a sub-directory with the
intention to just have a small include file in the root. Leaving that up
to the user to do until we can test it more with real dev flows.

With this, I was able to get:

* Minimal dev down to 3800 ninja actions to build iree-compile.
* ~5900 ninja actions for Turbine with a host-only CPU backend.

There are on the order of several hundred C++ files that I expect will
be eliminated once more of the compiler components are turned into
plugins with their dependent pieces contained (i.e. as opposed to having
dependencies for all targets implicitly included in top-level
registrations). Currently, this includes common code needed for SPIRV,
CPU compilation, various GPUs, etc (even when the corresponding target
is not enabled).

Further, I will do more surgery to make the clang build optional for
people who can satisfy the dep with a host clang. This will drop true
min-spec CPU builds by ~1000 targets.

Progress on #5804

skip-ci: none