Add vector_ext dialect (#15599)
This patch introduces the vector_ext dialect. The purpose of this
dialect is to have a place for experimenting with things beyond what the
upstream vector dialect provides.

In this particular PR, two new features are added
1. An explicit IR representation of the high dimensional layout that
looks like this #iree_vector_ext.per_dim_layout<"BatchX"<"LaneX"<"VecY",
2>, 4>, 4> The nesting makes clear what the innermost dimensions are and
their corresponding shapes.

2. Adds a layout conflict resolution operator. During layout analysis,
this operator can be used to resolve any differences in layout. The
lowering of this operator is not provided but the semantics are that
given a vector with an existing layout and a desired layout, the
operator transforms the vector to the desired layout.