[spirv] Use `gpu.subgrup_reduce` in SPIR-V vector reduction pipeline
This reduces the total number of supgroup ops from ~6 (with
`gpu.shuffle xor`) to 1 for supported full-warp reductions.

As an alternative to the new pass constructor argument, I considered
always emiting `gpu.subgroup_reduce` and later expanding it on targets
that do not support it (e.g., CUDA, ROCM). This was, however, more
difficult because of necessary type casts (required by `gpu.shuffle`),
and duplicated code accross the expansion and odd-sized warp reductions.

It may be possible replace some more shuffles with subgroup reductions,
but this can be a follow-up change.