Avoid fusing transpose linalg op with pack op (#15594)
Currently in `FormDispatchRegions` pass we use
`hasCompatibleOuterParallelLoops` to make sure the fused ops have
identity mapping between them on the parallel dimensions. The CPU
backend uses this assumption on tile size selections and tile-and-fuse.

This change adds the missing check to the special fusion logic on pack
ops, to ensure we don't fuse transpose generic op with pack op, which
results in the #15574 in CPU backend.

This should be considered as a workaround and eventually we might want
to support non-identitiy map fusion in CPU backend to remove these
constraints.