Adding support for `#util.uninitialized` globals. (#15657)
This allows global tensors to be initialized with `tensor.empty`
semantics useful for model state that will be overwritten by the model
itself prior to any reads. HAL implementations are still allowed to
initialize the memory but use of the attr in the compiler indicates the
program doesn't require them to do so.