Rationalize mmt4d tile sizes (#15651)
This builds on @hanhanW 's #15448 and further rationalizes mmt4d tile
sizes, to the extent that can be done without deeper changes to what's
determined at compile-time and what's treated as dynamic.
* Deduplicate KernelDispatch logic for `batch_mmt4d` vs `mmt4d`.
* Introduce a tile-size command-line option, `clGeneralMatmulTileBytes`
that's expressed in bytes (that a tile should fit in) as opposed to a
size along any particular dimension. This allows starting setting it to
values that can be heuristically justified, e.g. by comparision with the
size of some CPU cache, or with the amount of time that handling such a
tile would take vs. dispatch overhead.
* Introduce a separate tile-size option for narrow matmuls,
`clNarrowMatmulTileBytes`, to be used for `vecmat`, `matvec`, and
generally any matmul/batch_matmul where one of the sides is only one
data-tile tall. Indeed, the heuristic justification for a tile value is
very different in these narrow cases where the wide opposite operand is
only traversed once, so that the tile size can be set without
data-locality considerations.
* Remove unused experimental old mmt4d-specific command-line options.