[GlobalOpt] Do not set encodings on cast ops when the source is i1 type.
An input program could cast an i1 tensor to f32 tensor and use it in
matmul. We do not set encodings on such cast ops because mixing boalean
concept and i1 subtype concept is ambigious. We still want to data-tile
f32.f32.f32 matmul, but not include the cast op from i1 types.